refactor: extract reusable header component
